requirement.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
Django==4.0.1
django-cors-headers==3.11.0
django-filter==21.1
django-unixdatetimefield==1.2.0
djangorestframework==3.13.1
djangorestframework-queryfields==1.0.0
psycopg2==2.9.3
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
wincertstore==0.2
